Did a WHOLE bunch of stuff. Added: Spawn managers (July 11), Music/Sound (July 10), Cutscenes (July 12), Storyline stitching is possible but needs more content to test (July 14), Time travelling is completed (July 18), Structures can be created, loaded, and placed in the world (July 14). Also changed up a little bit of stuff within the entity manager, namely changed the way Entities are stored/passed around. Dealt with a lot of segfaults (kill me)

+++ Code/Console.h
@@ -29,7 +29,7 @@ class Console
 
         std::vector<CEGUI::DefaultWindow*> m_historyLabels;
         std::vector<std::string> m_commandHistory;
-        int m_historySelection = -1;
+        unsigned int m_historySelection = (unsigned int)-1;
 
         CEGUI::Editbox* m_editbox = nullptr;
